Restructure code to allow score computation on the host
Since OpenCL doesn't provide vector operations for use on the various cl_
types, a previous change removed support for native hand evaluation on
the host.  As a consequence, we had to make two kernel invocations per
compare_cards_opencl call: one to do most of the work and one to fill in
the last few entries.  This second call turns out to be incredibly expensive.

To get around this, I moved most of the score.cl code into score.h, and
added typedefs so the host gets a nonvectorized version of score_hands.
This change speeds up exact all from 6:02 s to 5:09 s.