Discovered that terminal output is a nontrivial performance hit
For some reason, writing even a moderate number of lines to the terminal
is a significant performance hit.  For example, piping the output of
exact some 100 to a file speeds it up from 4.3460 s to 3.042s.  With
output on, my timing code claims its spending 1.3250 s reading back
results, which decreases to 0.1200 s piped to a file.  If I disable
the read, the extra time switches to write free or compute, so it's
clearly not actually the read's fault.

Piping the output to a file and skipping tee speeds up exact all from
5:09 s to 4:51 s.