Add fixed precision rational number support to numpy
The "rational" module adds a new "rational" dtype to numpy stored as the
ratio of two 64-bit integers.  Overflow exceptions are thrown if any
inexactness is detected.  This allows us to take advantage of exact
arithmetic (and the exact probabilities provided by the "exact" program)
without sacrificing the convenience and speed of numpy.